chore: refactoring video adding stuff
